+++ packages/methods/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v1.0.1 (September 07, 2026)
 
+- Change `@formisch/core` to v1.0.1
 - Fix full form reset to discard pending validation results and clear the validating state (issue #210, pull request #211)
 - Fix field resets and array mutations to discard pending validation results for previous input
 - Fix submissions to ignore validation results superseded by input changes, full resets, or newer submissions